+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   [dev] Moves native view controllers to use iOS7+ status bar api - ash
 -   [dev] Renamed files to reflect case in component/function names - alloy
+-   Adds a media preview controller that can show remote document files - alloy
 
 ###### Emission